Add importer and scheduler

+++ sales/app/Console/Commands/importproducts.php
@@ -40,22 +40,28 @@ class importproducts extends Command
     public function handle()
     {
         $files  = Storage::files('imports');
-        foreach($files as $f){
-            $contents = Storage::get($f);
-            $contents = explode("\r\n", $contents);
-            $keys = str_getcsv($contents[0]);
+        if(count($files) > 0){
+            foreach($files as $f){
+                $contents = Storage::get($f);
+                $contents = explode("\r\n", $contents);
+                $keys = str_getcsv($contents[0]);
 
-            for ($i = 1; $i < count($contents); $i++) {
-                $values = str_getcsv($contents[$i]);
-                $values = array_combine($keys, $values);
+                for ($i = 1; $i < count($contents); $i++) {
+                    $values = str_getcsv($contents[$i]);
+                    $values = array_combine($keys, $values);
 
-                $prod = new Product($values);
-                $prod->image_url = '';
-                $prod->image_name = '';
+                    $prod = new Product($values);
+                    $prod->image_url = '';
+                    $prod->image_name = '';
 
-                $prod->save();
+                    $prod->save();
+                }
             }
+
+            print("Your import was completed.\n");
+            Storage::delete($files);
+        } else {
+            print("No file to be imported.\n");
         }
-        Storage::delete($files);
     }
 }
